backend/requirements.txt
mssql-django
django==4.2.11
djangorestframework
djangorestframework-simplejwt
mssql-django
celery>=5.0.0
redis>=4.0.0
django-cors-headers==4.7.0
Pillow>=10.0.0
django-filter>=2.4.0
faker>=20.0.0
drf-yasg
